[data-table] fixed render grouped header
