+++ semcore/data-table/src/components/Head/Group.tsx
@@ -34,26 +34,11 @@ export class Group extends Component<
   render() {
     const SGroupContainer = Box;
     const SGroup = Root;
-    const { styles, Children, title, fixed, columns, withConfig, getFixedStyle, shadowVertical } = this.asProps;
+    const { styles, Children, title, columns, withConfig, shadowVertical } = this.asProps;
     const groupColumns = columns ?? [];
 
     const firstColumn = groupColumns[0];
     const lastColumn = groupColumns[groupColumns.length - 1];
-
-    const style: any = {};
-
-    if (fixed === 'left' && firstColumn) {
-      const [name, value] = getFixedStyle({ name: firstColumn.name, fixed: 'left' });
-      if (name !== undefined && value !== undefined) {
-        style[name] = value;
-      }
-    }
-    if (fixed === 'right' && lastColumn) {
-      const [name, value] = getFixedStyle({ name: lastColumn.name, fixed: 'right' });
-      if (name !== undefined && value !== undefined) {
-        style[name] = value;
-      }
-    }
 
     return sstyled(styles)(
       <SGroupContainer data-group-container>
