+++ semcore/data-table/src/components/Head/Head.tsx
@@ -32,8 +32,31 @@ class HeadRoot<
     return `${uid}-column-sortable-describer`;
   }
 
-  getGroupProps(_: any, index: number) {
-    const { use, gridAreaGroupMap, children, getFixedStyle, shadowVertical } = this.asProps;
+  getGroupProps(props: any, index: number) {
+    const { fixed } = props;
+    const { use, gridAreaGroupMap, children, columns, getFixedStyle, shadowVertical, top, scrollDirection } = this.asProps;
+    const groupColumns = columns ?? [];
+
+    const firstColumn = groupColumns[0];
+    const lastColumn = groupColumns[groupColumns.length - 1];
+
+    const style: any = {};
+
+    if (fixed === 'left' && firstColumn) {
+      const [name, value] = getFixedStyle({ name: firstColumn.name, fixed: 'left' });
+      if (name !== undefined && value !== undefined) {
+        style[name] = value;
+      }
+    }
+    if (fixed === 'right' && lastColumn) {
+      const [name, value] = getFixedStyle({ name: lastColumn.name, fixed: 'right' });
+      if (name !== undefined && value !== undefined) {
+        style[name] = value;
+      }
+    }
+    if (top && scrollDirection !== 'horizontal') {
+      style.top = `${top}px`;
+    }
 
     return {
       use,
@@ -41,6 +64,8 @@ class HeadRoot<
       withConfig: children === undefined,
       getFixedStyle,
       shadowVertical,
+      style,
+      scrollDirection,
     };
   }
